DPC-4525 Github Actions Promote Build (#2443)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4525

## 🛠 Changes

- promote-build.yml added
- ecs-deploy.yml fix

## ℹ️ Context

We are migrating from Jenkins to Github Actions. This workflow replaces
'[DPC - Promote
Build](https://github.com/CMSgov/dpc-ops/blob/main/jenkins_files/Jenkinsfile.promote_build)'.

Note: Jenkins has a feature in which you can look up the parameters of
other jobs. This feature is not available in Github Actions. Therefore,
this promotion always uses the main branch of dpc-ops. We _could_ add
this as a parameter, but I think the cleaner interface is better.

## 🧪 Validation


[Ran](https://github.com/CMSgov/dpc-app/actions/runs/13317487378/job/37194959257)
and verified correct image and app_version deployed.