GHA Promote Fixes (#2483)
## 🎫 Ticket

No ticket

## 🛠 Changes

- role-to-assume updated for prod-sbx
- promote GHA pass input.to_env to Smoke Test
- ecs-deploy new relic notification logic updated

## ℹ️ Context

The first use of GHA for full deployment ran into the following issues:

- When assuming a role in AWS, the prod-sbx role is 'sbx', not
'prod-sbx', so this needs to be accounted for. Fix already in
smoke-test.yml
- promote-build.yml was passing the wrong input variable name to smoke
test, so it was testing the wrong environment.
- ecs-deploy.yml: new relic apm application search returns fuzzy
matches, so we need to make sure we get an application id as well. For
example, when you ask for "dpc-web-prod", it returns stanzas for
"dpc-web-prod-sbx", "dpc-web-admin-prod-sbx", and
"dpc-web-portal-prod-sbx", but nothing for "dpc-web-prod"

## 🧪 Validation

[Promote to
sandbox](https://github.com/CMSgov/dpc-app/actions/runs/13572910396)
worked (although not tested with smoke test).
[Promote from dev to
test](https://github.com/CMSgov/dpc-app/actions/runs/13573336062) smoke
tested dev.
Ran check locally and it blocked dpc-web-prod but passed dpc-web-test