Me/dpc 4545 fix aggregation (#2485)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4545

## 🛠 Changes

- Updated `completePartialBatch` so that it updates the job_queue_batch
table after each patient is processed.
- Temporarily Increased the job timeout to 15 minutes.

## ℹ️ Context

We noticed that jobs were getting stuck in the queue because the
update_time in the job_queue_batch table wasn't getting set frequently
enough. Sometimes it's because of issues in our DB code and sometimes
it's because patients are taking a long time to process. This should
clean up both of those.

## 🧪 Validation

- Added a test to ensure update_time is set after each patient is
processed.
- Deployed to test env and watched the DB as the smoke tests ran.

I verified that the code does what I expected it to, but the only way to
know for sure if this fixes our problem is to deploy it to prod and
periodically check the logs to see if jobs are still getting stuck.