[DPC-4632] Average Load Test (#2594)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4632

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
- Creates the configuration for the average load test
- Rewrites the workflow code to accommodate precise number of API calls
- Removes workflowB
- Added a single-iteration config so that we can test exactly one
run-through of the workflow

This PR consists of a pretty significant rewrite to the workflow logic.
`http.batch` is used in numerous places to run several API calls in
parallel.

Code is written to make the following number of API calls over the
course of the average load test:

- POST Patient - 700 total - 14 per iteration
- POST Practitioner - 50 total - 1 per iteration
- POST Group - 50 total - 1 per iteration
- GET Practitioner - 500 total - 10 per iteration
- GET Patient - 250 total - 5 per iteration
- GET Group - 50 total - 1 per iteration
- PUT Group - 250 total - 5 per iteration
- GET Group/$export - 250 total - 5 per iteration
- GET Job - 500 total - 10 per iteration

Over the course of the 50 iterations specified in the average-load-test
config, we'll hit the correct total number for each endpoint. This will
then scale for more iterations per hour when we configure the stress
test. Numbers are slightly adjusted from those in the ticket to simplify
counting. They all fall well within the ranges noted in the [Average
Hourly Load doc](confluence.cms.gov/display/DAPC/Hourly+API+Load).

In order to simplify the API call counts, `workflowB` was removed in
favor of a single workflow.

## ℹ️ Context

This enables us to run an Average Load Test, so that we can analyze how
our system operates under the average hourly load, in a controlled
environment.
<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
- [x] linter and tests pass
- [x] single-iteration run passes without failed http requests -
https://github.com/CMSgov/dpc-app/actions/runs/14598618165/job/40950837809