[DPC-4633] Configure Stress Test (#2604)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4633
## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
- Adds a configuration object for stress tests
- Sets a maxVUs number for stress tests
- Creates a new Makefile command to run stress tests locally

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
This allows us to run stress tests as part of our load testing suite.
Stress tests are designed to ramp up the load to test the limits of an
API -- in our case, we'll ramp up from 1x of the average load, to 2x,
3x, and finally 5x, in half hour increments.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->