Me/dpc 4659 fix aggregation mdc (#2608)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4659

## 🛠 Changes

Resets MDC between batches in dpc-aggregation.

## ℹ️ Context

Previously, when an aggregation node would log that it restarted a stuck
batch the MDC would still be filled in with information from the
previous batch (job_id, batch_id, etc...). This is because stuck batches
are handled between processed job batches, and the MDC was only getting
cleared when a new batch was started. Now it gets cleared after each
batch is finished.

## 🧪 Validation

Add MDC check to aggregation unit tests.