[DPC-4689] docker build workflow for greenfield environments (#2618)
## 🎫 Ticket

[DPC-4689](https://jira.cms.gov/browse/DPC-4689)

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
- update existing `pom.xml` and `docker-compose.yml` to remove hardcoded
account id
- update existing docker build workflow to take into account new changes
that don't have hardcoded account id
- added new workflow to perform docker build action for 2 new greenfield
aws accounts
 - fixed a couple broken readme links related to docker

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->
- For the greenfield migration, there are two new aws accounts (prod and
non-prod)
- After some discussion with folks in #dasg-greenfield, it seems the
most straightforward path to make docker images available in deployed
environments is to push to ECR on both accounts
- This ensures whatever is tagged `"latest"` is latest and we can
promote builds across accounts
- For the naming, we previously had the account id included as part of
the local image name that was on the CI runner before pushing. This is
because of the way it was defined for the jib-maven-plugin (see
pom.xml).
- This prefix has been removed from the locally created images, which
makes it easier to migrate to the new accounts. Also, the format is now
consistent between the rails apps and java apps.
- Expecting to remove original `docker-build.yml` and rename
`docker-build-gf.yml` accordingly when migration is complete

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
Confirmed that images were pushed to both accounts. See CI job
[HERE](https://github.com/CMSgov/dpc-app/actions/runs/14804448823)

---------

Co-authored-by: Michael J Burling <michael.burling@cms.hhs.gov>