[NO TICKET] disable AttributionFHIRTest (#2627)
## 🎫 Ticket

N/A

## 🛠 Changes

- disables AttributionFHIRTest

<!-- What was added, updated, or removed in this PR? -->

## ℹ️ Context

There was an attempt to resolve this ticket with
https://github.com/CMSgov/dpc-app/pull/2577. This appears to have not
been the solution, and now we're seeing an increased failure rate. This
is blocking PRs and builds on main, which is killing our velocity at a
time when we have the looming Greenfield cutover. Let's disable this for
now and revisit once the pressure's off.


<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->