[NO TICKET] pin ruby gem rack-session to avoid vulnerability (#2629)
## 🎫 Ticket
No ticket.
One of the Ruby gems is out of date and blocking the docker_build
workflow. See related job
[HERE](https://github.com/CMSgov/dpc-app/actions/runs/14935192842/job/41960872070).

## 🛠 Changes
- add rack-session to Gemfile files in order to pin a new enough version
that doesn't contain vulnerability triggering a failure in the pipeline
 - update rack version to latest (also to avoid vulnerability)

<!-- What was added, updated, or removed in this PR? -->

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
See "DPC CI Workflow / Build and Test DPC Portal" CI pipeline for this
pull request