[DPC-4690] Create Build and Deploy (Greenfield) workflow (#2641)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4690

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
This creates the new Build and Deploy workflow for Greenfield

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
This allows us to deploy the DPC API, web, admin, and portal to
Greenfield. For now, only dev is viable.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

- [x] Workflow runs successfully:
https://github.com/CMSgov/dpc-app/actions/runs/15119882494

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->