[DPC-4691] Create Promote Build (Greenfield) workflow (#2651)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4691

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
This creates the Promote Build workflow for Greenfield.

## ℹ️ Context

This will allow us to do our normal dev promotion workflow: `dev` ->
`test` -> `prod` and `test` -> `sandbox`.
<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

- [ ] Can successfully promote to both prod and
[non-prod](https://github.com/CMSgov/dpc-app/actions/runs/15279880611)
environments. Note that the sandbox deploy failed at one of the [health
checks](https://github.com/CMSgov/dpc-app/actions/runs/15307219466/job/43063025383).
This is not an issue with this workflow, and given that it is able to
deploy, my take is that this workflow works as intended within its own
scope.
- [ ] After merging this PR, promote to test, then to sandbox and prod.

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->