[On Call] DropWizard 4.0.11 -> 4.0.14 (#2669)
## 🎫 Ticket

None, part of on call

## 🛠 Changes

- Update DropWizard from 4.0.11 -> 4.0.14
- Update location of `IAccessEvent` class in `AuthRelatedLoggingFilter`.
- Remove direct dependency on `dropwizard-json-logging`, as it's already
included indirectly in the dropwizard bom.
- Pinned `logback-access` and `logback-core`.

## ℹ️ Context

Snyk opened https://github.com/CMSgov/dpc-app/pull/2640 to upgrade
`dropwizard-json-logging`, and all of DropWizard, from 4.0.11 to 4.0.13.
The PR fails, though, because this version of drop wizard tries to pull
in a new version of logback while drop wizard-jobs tries to pull in the
old one, and because the `IAccessEvent` class was moved to a new
location in the logback library.

## 🧪 Validation

Program compiles and all tests pass.