[DPC-4755] migrate remaining GHA workflows to run on codebuild (#2686)
## 🎫 Ticket

[DPC-4755](https://jira.cms.gov/browse/DPC-4755)

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
This PR aims to update the `runs-on` property for workflows that are not
already set up to run on AWS codebuild.
I found 4 workflow files that needed to be moved over.
 - generate macaroon
 - axe accessibility tests
 - opt-out import unit tests
 - tag release

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->
As part of the migration to new AWS greenfield accounts, the team has
already modified existing workflows essential to building and deploying
the system. This PR moves remaining workflows to AWS codebuild, which
runs in the new account. This eliminates the need to have github runners
for dpc-app running in the legacy AWS account.


## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

**Macaroons here:**

https://github.com/CMSgov/dpc-app/actions/runs/15829941153/job/44619818120

**Accessibility tests:**
Axe tests run kick off here, but still fail (outside of scope):

https://github.com/CMSgov/dpc-app/actions/runs/15829479118/job/44618231030#logs

**Opt-out-import unit tests:**
Go-version is not specified
Last successful run was on version go1.24.4
Job HERE:
https://github.com/CMSgov/dpc-app/actions/runs/15825619747/job/44604920355
Successful run here:
https://github.com/CMSgov/dpc-app/actions/runs/15830355906

**Tag release**
See modified workflow run here with tag output to stdout

https://github.com/CMSgov/dpc-app/actions/runs/15835252904/job/44637231615