Me/dpc 4710 graceful aggregation shutdown v2 (#2706)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4710

## 🛠 Changes

- When `stop()` is called on the aggregation engine it, immediately
pauses a currently running batch instead of waiting for the current
patient to finish processing and then pausing the batch.
- Adds `TEST_PATIENT_TIME_OUT` mbi to the MockBlueButtonClient. It's an
mbi that when requested will cause the client to block and wait forever
and is used to simulate long running BFD requests.
- Updates dpc-api to use the `MockBfdClient` instead of the real BFD
sandbox when running locally. This was already the case for
dpc-aggregation, and since we can't connect to BFD from our local
machines anymore this was preventing us from testing.

## ℹ️ Context

We were seeing regular alerts about stuck batches getting restarted
whenever dpc-aggregation scaled down and nodes were getting removed.

The way things were supposed to work:
1. AWS sends a SIGINT to a dpc-aggregation node telling it to shut down.
2. A Jetty shutdown thread is started, and it calls the
`AggregationEngine.stop()` method.
3. `stop()` sets the AggregationEngine's queue to no longer running.
4. If a batch is currently being processed, the current patient is
finished and then the batch is paused (status set to QUEUED).
5. The shutdown thread finishes cleaning up other resources that
implement the `Managed` interface and the node shuts down.
6. If the node hasn't shut down 30 seconds after the SIGINT, a SIGKILL
is sent and it's removed.

The problem is happening at step 4. We often have patients that take
multiple minutes to process. If one of those patients is running when
the SIGINT is sent, then there isn't time for it to finish and the batch
to be paused before the aggregation node is killed. The batch is left in
a running state on the queue, and after 15 minutes of no activity,
another node restarts it and logs a "Restarting stuck batch..." warning.
If we get enough of these in a given time period an alert is sent to
#dpc-alerts.

This PR moves the call to `pauseBatch()` into the engine's `stop()`
method. This means it's called in the Jetty shutdown thread and happens
immediately, while the current patient is still being processed, instead
of waiting for it to finish.

## 🧪 Validation

- Deployed to dev and smoke tests passed (link
[here](https://github.com/CMSgov/dpc-app/actions/runs/16026107894/job/45218537918)).
- Created a new test that simulates shutting down a node while a patient
is being processed.
- Was able to shutdown a dpc-aggregation node normally (logs
[here](https://us-east-1.console.aws.amazon.com/cloudwatch/home?region=us-east-1#logsV2:log-groups/log-group/%2Faws%2Fecs%2Ffargate%2Fdpc-dev-aggregation/log-events/aggregation%2Fdpc-dev-aggregation%2F636ee76219ae4c0182c840ee71b917b6)).

## Note:
This fixes the problem described above which seems to be causing
most/all of our stuck batch issues, but not every problem. There may
still occasionally be stuck batch warnings that need to be investigated.

Failing SonarQube test being addressed in
https://github.com/CMSgov/dpc-app/pull/2705