[dpc-4710] Shuts down dpc-queue after AggregationEngine (#2708)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4710
Bugfix for https://github.com/CMSgov/dpc-app/pull/2706

## 🛠 Changes

Updated the order that `DPCQueueHibernateModule` and
`AggregationAppModule` are added to the `DPCAggregationService`.

## ℹ️ Context

DropWizard shuts these modules down in the opposite order they are
added, and we don't want the queue to shutdown until after the
aggregation engine has finished processing or it won't be able to update
its last batch.

This wasn't caught in initial testing for two reasons:

1. Our tests used the `MemoryBatchQueue` which doesn't connect to an
external DB.
2. Our code was ignoring the errors when it tried to pause a batch
instead of logging them.

## 🧪 Validation

1. Go into your local queue DB and pick out your most recently completed
batch.
2. Update the batch's `status` to 0 (_queued_), its `aggregator_id` to
`null`, and its patient to `9S99EU8XY92`. (this patient tells
MockBfdClient to hang forever, simulating a long running patient)
3. Start up dpc-aggregation and it will pick up the batch and set its
status to 1 (_running_).
4. Stop dpc-aggregation.  

If you do this on the current main branch, the status of the batch will
still be 1 (_running_). If you recompile from this branch and go through
the same process the batch will be reset to 0 (_queued_), which is what
we want.