PLT-657:Rename the public platform github repo (#2595)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-657

## 🛠 Changes

references to ab2d-bcda-dpc-platform renamed to cdap

## ℹ️ Context

we are renaming the ab2d-bcda-dpc-platform repo to cdap

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation


upon approval, ab2d-bcda-dpc-platform repo will be renamed to cdap and
workflow re-ran, if successful PR merged if not ab2d-bcda-dpc-platform
repo rename to cdap reverted while investigating why workflow failed

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>