DPC-4790: Call resources only once (#2736)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4790

## 🛠 Changes

- ResourceFetcher returns bundles instead of flowable closure
- JobBatchProcessor wraps the bundles in a flowable instead of just
returning the results from ResourceFetcher
- ResourceFetcherUnitTests: changes to handle signature change.

## ℹ️ Context

We would like to call BFD only one time, and the code _looks_ like that
is what it does. But it turns out we were returning a closure that
called BFD, so it was still calling it. This code just returns the
results of the calls from BFD.

## 🧪 Validation

Did export of data from TEST_PATIENT[2] from MockBlueButtonClient in
both main and this branch. The data returned from the export was the
same (except for order), but the aggregation logs for
dpcMetric:ResourceFetched showed two calls
resourceType:ExplanationOfBenefit when running in main, but only one
call in this branch.