Update all package-lock.json files (#2737)
## 🎫 Ticket
No ticket

## 🛠 Changes

- docker-compose.portals.yml updated to make package-lock.json available
on local system
- package-lock.json updates for dpc-admin, dpc-web, dpc-portal

## ℹ️ Context

Our security scans show issues in our dpc-web and dpc-portal
package-lock.json files, but these files were not up-to-date, so they
needed to be updated. dpc-admin has already been updated to avoid
security concerns, but it seemed easier to update it again.

As before, the package.json fail in snyk is unrelated, and also a little
odd, as the library it references is already at the version it requests.

Updated by
- starting portals with `make start-portals`
- in each portal
```
apk add npm
rm -rf node_modules
npm install
```

## 🧪 Validation

All of the package-lock.json files showed updates, and the checks pass.