Portal works even if ALLOW_INVALID_SSL_CERT not set (#2746)
## 🎫 Ticket

No ticket

## 🛠 Changes

Updated fetch of ALLOW_INVALID_SSL_CERT to be false unless set

## ℹ️ Context

Running local was failing because ALLOW_INVALID_SSL_CERT was not set. We
could always force it to be set, or assume false if it wasn't set. The
latter made more sense to me.

## 🧪 Validation

Ran dpc-portal locally with the new client and successfully added a
public key.