[DPC-4812] Moved foreman to Docker image (#2747)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4812

## 🛠 Changes

Moved Foreman gem into the portal Docker image.

## ℹ️ Context

We had been installing Foreman on a per environment basis when the
portal starts up, but as part of
https://github.com/CMSgov/dpc-ops/pull/853 we won't be able to write to
our containers anymore. That means we can no longer install anything, so
we're adding Foreman to the Docker image ahead of time instead. The
trade off is it'll get installed regardless of whether we actually
intend to use it, but I don't think that's a big deal.

## 🧪 Validation

Full build and deploy with passing smoke tests
[https://github.com/CMSgov/dpc-app/actions/runs/16526473477](https://github.com/CMSgov/dpc-app/compare/main...me/here).