Me/dpc 4817 use iam for db access (#2756)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4817

## 🛠 Changes

- In the test environment only, use IAM authentication to get to the DB
in dpc-api.
- Remove old prod-sbx env files.

## ℹ️ Context

For security reasons, we're switching to using IAM auth to access the
new RDS. This is currently setup in the test environment, but not the
others, so they still need to use the older method.

### Note:

When we're ready to port the DB changes to our other environments, it
should just be a matter of updating each services ENV.application.env
file.

## 🧪 Validation

Full deploy and smoke test to dev (old auth):
https://github.com/CMSgov/dpc-app/actions/runs/16914573407

Full deploy and smoke test to test (new auth):
https://github.com/CMSgov/dpc-app/actions/runs/16914587330 (Smoke test
passed the
[second](https://github.com/CMSgov/dpc-app/actions/runs/16916276852)
time)

Ran full build locally and all services can still connect to local DB.