Me/dpc 4875 fix patient consent tests (#2787)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4875

## 🛠 Changes

- Fix patient integration tests.
- Updated `change-type-consent-effective-date` migration to run when
column is not a `date` as opposed to when it doesn't exist.

## ℹ️ Context

With dpc-consent gone and the consent DB living in dpc-aggregation,
integration tests that relied on setting up opt outs no longer work.
This fixes them by writing directly to the consent DB instead of the
soon to be non-existent dpc-consent service.

## 🧪 Validation

All tests enabled and passing.

---------

Co-authored-by: Ashley Weaver <ashleyweaver@navapbc.com>
Co-authored-by: Ashley Weaver <134093673+ashley-weaver@users.noreply.github.com>
Co-authored-by: jdettmannnava <145699825+jdettmannnava@users.noreply.github.com>