DPC-4913 build xk6 (#2796)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4913

## 🛠 Changes

xk6 built from source code rather than via docker.

## ℹ️ Context

Because github actions run themselves in containers, we want to reduce
the number of docker containers we run in them.

I was unable to remove the statsd container that connects to New Relic.
The docker container for this build uses another statsd container, and I
was unable to build a raw copy. The statsd container that connects to
Data Dog suffers the same problem.

## 🧪 Validation

Successful load test:
https://github.com/CMSgov/dpc-app/actions/runs/18203336896